Removed form_profile_v2 spec
