+++ spec/models/form_profile_spec.rb
@@ -8,27 +8,26 @@ require 'gi/client'
 RSpec.describe FormProfile, type: :model do
   include SchemaMatchers
 
-  let(:user) do
-    build(:user, :loa3, :legacy_icn, suffix: 'Jr.', idme_uuid: 'b2fab2b5-6af0-45e1-a9e2-394347af91ef',
-                                     address: build(:mpi_profile_address))
-  end
-
   before do
-    stub_evss_pciu(user)
-    described_class.instance_variable_set(:@mappings, nil)
     allow(Flipper).to receive(:enabled?).and_call_original
-    allow(Flipper).to receive(:enabled?).with(:remove_pciu, anything).and_return(false)
+    allow(Flipper).to receive(:enabled?).with(:remove_pciu, instance_of(User)).and_return(true)
+    allow(Flipper).to receive(:enabled?).with(:remove_pciu).and_return(true)
     allow(Flipper).to receive(:enabled?).with(:disability_526_max_cfi_service_switch, anything).and_return(false)
+    described_class.instance_variable_set(:@mappings, nil)
   end
 
-  let(:street_check) { build(:street_check) }
+  let(:user) do
+    build(:user, :loa3, :legacy_icn, idme_uuid: 'b2fab2b5-6af0-45e1-a9e2-394347af91ef', suffix: 'Jr.',
+                                     address: build(:va_profile_v3_address), vet360_id: '1')
+  end
 
+  let(:contact_info) { form_profile.send :initialize_contact_information }
   let(:form_profile) do
     described_class.new(form_id: 'foo', user:)
   end
-
-  let(:us_phone) { form_profile.send :pciu_us_phone }
-
+  let(:va_profile_address) { contact_info&.address }
+  let(:us_phone) { contact_info&.home_phone }
+  let(:mobile_phone) { contact_info&.mobile_phone }
   let(:full_name) do
     {
       'first' => user.first_name&.capitalize,
@@ -37,7 +36,6 @@ RSpec.describe FormProfile, type: :model do
       'suffix' => user.suffix
     }
   end
-
   let(:veteran_service_information) do
     {
 
@@ -48,30 +46,26 @@ RSpec.describe FormProfile, type: :model do
       }
     }
   end
-
   let(:veteran_full_name) do
     {
       'veteranFullName' => full_name
     }
   end
-
   let(:address) do
     {
-      'street' => street_check[:street],
-      'street2' => street_check[:street2],
-      'city' => user.address[:city],
-      'state' => user.address[:state],
-      'country' => user.address[:country],
-      'postal_code' => user.address[:postal_code].slice(0, 5)
+      'street' => va_profile_address.street,
+      'street2' => va_profile_address.street2,
+      'city' => va_profile_address.city,
+      'state' => va_profile_address.state,
+      'country' => va_profile_address.country,
+      'postal_code' => va_profile_address.postal_code
     }
   end
-
   let(:veteran_address) do
     {
       'veteranAddress' => address
     }
   end
-
   let(:tours_of_duty) do
     [
       {
@@ -96,7 +90,6 @@ RSpec.describe FormProfile, type: :model do
       }
     ]
   end
-
   let(:v40_10007_expected) do
     {
       'application' => {
@@ -105,13 +98,12 @@ RSpec.describe FormProfile, type: :model do
           'dateOfBirth' => user.birth_date,
           'name' => full_name,
           'ssn' => user.ssn,
-          'email' => user.pciu_email,
+          'email' => user.va_profile_email,
           'phoneNumber' => us_phone
         }
       }
     }
   end
-
   let(:v0873_expected) do
     user_work_phone = user.vet360_contact_info.work_phone
     work_phone = [
@@ -133,7 +125,7 @@ RSpec.describe FormProfile, type: :model do
         'serviceNumber' => '123455678'
       },
       'contactInformation' => {
-        'email' => user.pciu_email,
+        'email' => user.va_profile_email,
         'phone' => us_phone,
         'address' => address,
         'workPhone' => work_phone
@@ -149,7 +141,6 @@ RSpec.describe FormProfile, type: :model do
       'veteranServiceInformation' => veteran_service_information
     }
   end
-
   let(:v686_c_674_expected) do
     {
       'veteranContactInformation' => {
@@ -160,8 +151,8 @@ RSpec.describe FormProfile, type: :model do
           'stateCode' => 'DC',
           'zipCode' => '20011'
         },
-        'phoneNumber' => '4445551212',
-        'emailAddress' => 'test2@test1.net'
+        'phoneNumber' => '3035551234',
+        'emailAddress' => user.va_profile_email
       },
       'nonPrefill' => {
         'veteranSsnLastFour' => '1863',
@@ -189,8 +180,8 @@ RSpec.describe FormProfile, type: :model do
           'stateCode' => 'DC',
           'zipCode' => '20011'
         },
-        'phoneNumber' => '4445551212',
-        'emailAddress' => 'test2@test1.net'
+        'phoneNumber' => us_phone,
+        'emailAddress' => user.va_profile_email
       },
       'nonPrefill' => {
         'veteranSsnLastFour' => '1863',
@@ -342,21 +333,13 @@ RSpec.describe FormProfile, type: :model do
       ]
     }
   end
-
   let(:v22_1990_expected) do
     {
       'toursOfDuty' => tours_of_duty,
       'currentlyActiveDuty' => {
         'yes' => false
       },
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'veteranFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -365,12 +348,12 @@ RSpec.describe FormProfile, type: :model do
       },
       'gender' => user.gender,
       'homePhone' => us_phone,
+      'mobilePhone' => mobile_phone,
       'veteranDateOfBirth' => user.birth_date,
       'veteranSocialSecurityNumber' => user.ssn,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email
     }
   end
-
   let(:v22_0993_expected) do
     {
       'claimantFullName' => {
@@ -382,18 +365,10 @@ RSpec.describe FormProfile, type: :model do
       'claimantSocialSecurityNumber' => user.ssn
     }
   end
-
   let(:v22_0994_expected) do
     {
       'activeDuty' => false,
-      'mailingAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'mailingAddress' => address,
       'applicantFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -402,26 +377,19 @@ RSpec.describe FormProfile, type: :model do
       },
       'applicantGender' => user.gender,
       'dayTimePhone' => us_phone,
+      'nightTimePhone' => mobile_phone,
       'dateOfBirth' => user.birth_date,
       'applicantSocialSecurityNumber' => user.ssn,
-      'emailAddress' => user.pciu_email
+      'emailAddress' => user.va_profile_email
     }
   end
-
   let(:v22_1990_n_expected) do
     {
       'toursOfDuty' => tours_of_duty,
       'currentlyActiveDuty' => {
         'yes' => false
       },
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'veteranFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -432,20 +400,12 @@ RSpec.describe FormProfile, type: :model do
       'homePhone' => us_phone,
       'veteranDateOfBirth' => user.birth_date,
       'veteranSocialSecurityNumber' => user.ssn,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email
     }
   end
-
   let(:v22_1990_e_expected) do
     {
-      'relativeAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'relativeAddress' => address,
       'relativeFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -455,17 +415,9 @@ RSpec.describe FormProfile, type: :model do
       'relativeSocialSecurityNumber' => user.ssn
     }
   end
-
   let(:v22_1995_expected) do
     {
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'veteranFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -474,20 +426,12 @@ RSpec.describe FormProfile, type: :model do
       },
       'homePhone' => us_phone,
       'veteranSocialSecurityNumber' => user.ssn,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email
     }
   end
-
   let(:v22_1995_s_expected) do
     {
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'veteranFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -496,20 +440,12 @@ RSpec.describe FormProfile, type: :model do
       },
       'homePhone' => us_phone,
       'veteranSocialSecurityNumber' => user.ssn,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email
     }
   end
-
   let(:v22_10203_expected) do
     {
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'veteranFullName' => {
         'first' => user.first_name&.capitalize,
         'middle' => user.middle_name&.capitalize,
@@ -517,11 +453,11 @@ RSpec.describe FormProfile, type: :model do
         'suffix' => user.suffix
       },
       'homePhone' => us_phone,
+      'mobilePhone' => mobile_phone,
       'veteranSocialSecurityNumber' => user.ssn,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email
     }
   end
-
   let(:v22_5490_expected) do
     {
       'toursOfDuty' => tours_of_duty,
@@ -536,7 +472,6 @@ RSpec.describe FormProfile, type: :model do
       'relativeDateOfBirth' => user.birth_date
     }
   end
-
   let(:v22_5495_expected) do
     {
       'toursOfDuty' => tours_of_duty,
@@ -551,7 +486,6 @@ RSpec.describe FormProfile, type: :model do
       'relativeDateOfBirth' => user.birth_date
     }
   end
-
   let(:v1010ez_expected) do
     {
       'veteranFullName' => {
@@ -561,15 +495,8 @@ RSpec.describe FormProfile, type: :model do
         'suffix' => user.suffix
       },
       'veteranDateOfBirth' => user.birth_date,
-      'email' => user.pciu_email,
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'email' => user.va_profile_email,
+      'veteranAddress' => address,
       'swAsiaCombat' => false,
       'lastServiceBranch' => 'army',
       'lastEntryDate' => '2002-07-02',
@@ -579,7 +506,6 @@ RSpec.describe FormProfile, type: :model do
       'veteranSocialSecurityNumber' => user.ssn
     }
   end
-
   let(:vmdot_expected) do
     {
       'fullName' => {
@@ -674,7 +600,6 @@ RSpec.describe FormProfile, type: :model do
       ]
     }
   end
-
   let(:v5655_expected) do
     {
       'personalIdentification' => {
@@ -685,7 +610,7 @@ RSpec.describe FormProfile, type: :model do
         'veteranFullName' => full_name,
         'address' => address,
         'telephoneNumber' => us_phone,
-        'emailAddress' => user.pciu_email,
+        'emailAddress' => user.va_profile_email,
         'dateOfBirth' => user.birth_date
       },
       'income' => [
@@ -696,10 +621,9 @@ RSpec.describe FormProfile, type: :model do
       ]
     }
   end
-
   let(:vvic_expected) do
     {
-      'email' => user.pciu_email,
+      'email' => user.va_profile_email,
       'serviceBranches' => ['F'],
       'gender' => user.gender,
       'verified' => true,
@@ -708,7 +632,6 @@ RSpec.describe FormProfile, type: :model do
       'veteranSocialSecurityNumber' => user.ssn
     }.merge(veteran_full_name).merge(veteran_address)
   end
-
   let(:v21_p_527_ez_expected) do
     {
       'veteranFullName' => {
@@ -717,28 +640,22 @@ RSpec.describe FormProfile, type: :model do
         'last' => user.last_name&.capitalize,
         'suffix' => user.suffix
       },
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
-      'email' => 'test2@test1.net',
-      'phone' => '4445551212',
-      'internationalPhone' => '14445551212',
-      'veteranSocialSecurityNumber' => user.ssn,
-      'veteranDateOfBirth' => user.birth_date,
-      'activeServiceDateRange' => {
-        'from' => '1984-08-01',
-        'to' => '2014-08-31'
-      },
+      'veteranAddress' => address,
+      'email' => user.va_profile_email,
+      'phone' => us_phone,
+      'mobilePhone' => mobile_phone,
+      'internationalPhone' => '3035551234',
       'serviceBranch' => {
         'army' => true,
         'airForce' => true
       },
-      'serviceNumber' => '796111863'
+      'activeServiceDateRange' => {
+        'from' => '1984-08-01',
+        'to' => '2014-08-31'
+      },
+      'serviceNumber' => '796111863',
+      'veteranSocialSecurityNumber' => user.ssn,
+      'veteranDateOfBirth' => user.birth_date
     }
   end
 
@@ -750,16 +667,9 @@ RSpec.describe FormProfile, type: :model do
         'last' => user.last_name&.capitalize,
         'suffix' => user.suffix
       },
-      'claimantAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => 'USA',
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'claimantAddress' => address,
       'claimantPhone' => us_phone,
-      'claimantEmail' => user.pciu_email
+      'claimantEmail' => user.va_profile_email
     }
   end
 
@@ -812,8 +722,8 @@ RSpec.describe FormProfile, type: :model do
           'zipCode' => '20011',
           'addressLine1' => '140 Rock Creek Rd'
         },
-        'primaryPhone' => '4445551212',
-        'emailAddress' => 'test2@test1.net'
+        'primaryPhone' => '3035551234',
+        'emailAddress' => 'person101@example.com'
       },
       'bankAccountNumber' => '******7890',
       'bankAccountType' => 'Checking',
@@ -823,16 +733,15 @@ RSpec.describe FormProfile, type: :model do
       'syncModern0781Flow' => true
     }
   end
-
   let(:vfeedback_tool_expected) do
     {
       'address' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
+        'street' => va_profile_address.street,
+        'street2' => va_profile_address.street2,
+        'city' => va_profile_address.city,
+        'state' => va_profile_address.state,
         'country' => 'US',
-        'postal_code' => user.address[:postal_code][0..4]
+        'postal_code' => va_profile_address.postal_code
       },
       'serviceBranch' => 'Army',
       'fullName' => {
@@ -841,7 +750,7 @@ RSpec.describe FormProfile, type: :model do
         'last' => user.last_name&.capitalize,
         'suffix' => user.suffix
       },
-      'applicantEmail' => user.pciu_email,
+      'applicantEmail' => user.va_profile_email,
       'phone' => us_phone,
       'serviceDateRange' => {
         'from' => '2002-07-02',
@@ -849,7 +758,6 @@ RSpec.describe FormProfile, type: :model do
       }
     }
   end
-
   let(:v26_1880_expected) do
     {
       'fullName' => {
@@ -859,16 +767,9 @@ RSpec.describe FormProfile, type: :model do
         'suffix' => user.suffix
       },
       'dateOfBirth' => '1809-02-12',
-      'applicantAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'applicantAddress' => address,
       'contactPhone' => us_phone,
-      'contactEmail' => user.pciu_email,
+      'contactEmail' => user.va_profile_email,
       'periodsOfService' => tours_of_duty,
       'currentlyActiveDuty' => {
         'yes' => false
@@ -876,22 +777,13 @@ RSpec.describe FormProfile, type: :model do
       'activeDuty' => false
     }
   end
-
   let(:v28_8832_expected) do
     {
-      'claimantAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'claimantAddress' => address,
       'claimantPhoneNumber' => us_phone,
-      'claimantEmailAddress' => user.pciu_email
+      'claimantEmailAddress' => user.va_profile_email
     }
   end
-
   let(:vform_mock_ae_design_patterns_expected) do
     {
       'data' => {
@@ -905,26 +797,26 @@ RSpec.describe FormProfile, type: :model do
             'ssn' => user.ssn.last(4),
             'gender' => user.gender,
             'address' => {
-              'addressLine1' => street_check[:street],
-              'addressLine2' => street_check[:street2],
-              'city' => user.address[:city],
-              'stateCode' => user.address[:state],
-              'countryName' => user.address[:country],
-              'zipCode5' => user.address[:postal_code][0..4]
+              'addressLine1' => va_profile_address.street,
+              'addressLine2' => va_profile_address.street2,
+              'city' => va_profile_address.city,
+              'stateCode' => va_profile_address.state,
+              'countryName' => va_profile_address.country,
+              'zipCode5' => va_profile_address.postal_code
             },
             'phone' => {
               'areaCode' => us_phone[0..2],
               'phoneNumber' => us_phone[3..9]
             },
-            'homePhone' => '14445551212',
-            'emailAddressText' => user.pciu_email,
+            'homePhone' => '3035551234',
+            'mobilePhone' => mobile_phone,
+            'emailAddressText' => user.va_profile_email,
             'lastServiceBranch' => 'Army'
           }
         }
       }
     }
   end
-
   let(:v28_1900_expected) do
     {
       'veteranInformation' => {
@@ -936,19 +828,12 @@ RSpec.describe FormProfile, type: :model do
         'ssn' => '796111863',
         'dob' => '1809-02-12'
       },
-      'veteranAddress' => {
-        'street' => street_check[:street],
-        'street2' => street_check[:street2],
-        'city' => user.address[:city],
-        'state' => user.address[:state],
-        'country' => user.address[:country],
-        'postal_code' => user.address[:postal_code][0..4]
-      },
+      'veteranAddress' => address,
       'mainPhone' => us_phone,
-      'email' => user.pciu_email
+      'email' => user.va_profile_email,
+      'cellPhone' => mobile_phone
     }
   end
-
   let(:v21_22_expected) do
     {
       'personalInformation' => {
@@ -961,16 +846,9 @@ RSpec.describe FormProfile, type: :model do
         'dateOfBirth' => '1809-02-12'
       },
       'contactInformation' => {
-        'email' => user.pciu_email,
-        'address' => {
-          'street' => street_check[:street],
-          'street2' => street_check[:street2],
-          'city' => user.address[:city],
-          'state' => user.address[:state],
-          'country' => user.address[:country],
-          'postal_code' => user.address[:postal_code][0..4]
-        },
-        'primaryPhone' => '4445551212'
+        'email' => user.va_profile_email,
+        'address' => address,
+        'primaryPhone' => '3035551234'
       },
       'militaryInformation' => {
         'serviceBranch' => 'Army',
@@ -986,7 +864,6 @@ RSpec.describe FormProfile, type: :model do
       }
     }
   end
-
   let(:v21_22_a_expected) do
     {
       'personalInformation' => {
@@ -999,16 +876,9 @@ RSpec.describe FormProfile, type: :model do
         'dateOfBirth' => '1809-02-12'
       },
       'contactInformation' => {
-        'email' => user.pciu_email,
-        'address' => {
-          'street' => street_check[:street],
-          'street2' => street_check[:street2],
-          'city' => user.address[:city],
-          'state' => user.address[:state],
-          'country' => user.address[:country],
-          'postal_code' => user.address[:postal_code][0..4]
-        },
-        'primaryPhone' => '4445551212'
+        'email' => user.va_profile_email,
+        'address' => address,
+        'primaryPhone' => '3035551234'
       },
       'militaryInformation' => {
         'serviceBranch' => 'Army',
@@ -1024,7 +894,6 @@ RSpec.describe FormProfile, type: :model do
       }
     }
   end
-
   let(:v26_4555_expected) do
     {
       'veteran' => {
@@ -1035,20 +904,13 @@ RSpec.describe FormProfile, type: :model do
         },
         'ssn' => '796111863',
         'dateOfBirth' => '1809-02-12',
-        'homePhone' => '14445551212',
-        'email' => user.pciu_email,
-        'address' => {
-          'street' => street_check[:street],
-          'street2' => street_check[:street2],
-          'city' => user.address[:city],
-          'state' => user.address[:state],
-          'country' => user.address[:country],
-          'postal_code' => user.address[:postal_code][0..4]
-        }
+        'homePhone' => us_phone,
+        'mobilePhone' => mobile_phone,
+        'email' => user.va_profile_email,
+        'address' => address
       }
     }
   end
-
   let(:v21_0966_expected) do
     {
       'veteran' => {
@@ -1059,16 +921,9 @@ RSpec.describe FormProfile, type: :model do
         },
         'ssn' => '796111863',
         'dateOfBirth' => '1809-02-12',
-        'homePhone' => '14445551212',
-        'email' => user.pciu_email,
-        'address' => {
-          'street' => street_check[:street],
-          'street2' => street_check[:street2],
-          'city' => user.address[:city],
-          'state' => user.address[:state],
-          'country' => user.address[:country],
-          'postal_code' => user.address[:postal_code][0..4]
-        }
+        'homePhone' => '13035551234',
+        'email' => user.va_profile_email,
+        'address' => address
       }
     }
   end
@@ -1083,8 +938,8 @@ RSpec.describe FormProfile, type: :model do
         },
         'ssn' => '1863',
         'dateOfBirth' => '1809-02-12',
-        'homePhone' => '14445551212',
-        'email' => user.pciu_email,
+        'homePhone' => us_phone,
+        'email' => user.va_profile_email,
         'fileNumber' => '3735'
       }
     }
@@ -1243,39 +1098,6 @@ RSpec.describe FormProfile, type: :model do
     end
   end
 
-  describe '#pciu_us_phone' do
-    def self.test_pciu_us_phone(primary, expected)
-      it "returns #{expected}" do
-        allow_any_instance_of(FormProfile).to receive(:pciu_primary_phone).and_return(primary)
-        expect(form_profile.send(:pciu_us_phone)).to eq(expected)
-      end
-    end
-
-    context 'with nil' do
-      test_pciu_us_phone(nil, '')
-    end
-
-    context 'with an intl phone number' do
-      test_pciu_us_phone('442079460976', '')
-    end
-
-    context 'with a us phone number' do
-      test_pciu_us_phone('5557940976', '5557940976')
-    end
-
-    context 'with a us 1+ phone number' do
-      test_pciu_us_phone('15557940976', '5557940976')
-    end
-  end
-
-  describe '#extract_pciu_data' do
-    it 'rescues EVSS::ErrorMiddleware::EVSSError errors' do
-      expect(user).to receive(:pciu_primary_phone).and_raise(EVSS::ErrorMiddleware::EVSSError)
-
-      expect(form_profile.send(:extract_pciu_data, :pciu_primary_phone)).to eq('')
-    end
-  end
-
   describe '#prefill_form' do
     def can_prefill_vaprofile(yes)
       expect(user).to receive(:authorize).at_least(:once).with(:va_profile, :access?).and_return(yes)
@@ -1306,7 +1128,6 @@ RSpec.describe FormProfile, type: :model do
       end.tap do |schema_form_id|
         schema = strip_required(VetsJsonSchema::SCHEMAS[schema_form_id]).except('anyOf')
         schema_data = prefilled_data.deep_dup
-
         errors = JSON::Validator.fully_validate(
           schema,
           schema_data.deep_transform_keys { |key| key.camelize(:lower) }, validate_schema: true
@@ -1324,7 +1145,6 @@ RSpec.describe FormProfile, type: :model do
       let(:form_profile) do
         FormProfiles::VA1010ezr.new(user:, form_id: 'f')
       end
-
       let(:ezr_prefilled_data_without_ee_data) do
         {
           'veteranFullName' => {
@@ -1337,15 +1157,8 @@ RSpec.describe FormProfile, type: :model do
           'gender' => user.gender,
           'veteranDateOfBirth' => user.birth_date,
           'homePhone' => us_phone,
-          'veteranAddress' => {
-            'street' => street_check[:street],
-            'street2' => street_check[:street2],
-            'city' => user.address[:city],
-            'state' => user.address[:state],
-            'country' => user.address[:country],
-            'postal_code' => user.address[:postal_code][0..4]
-          },
-          'email' => user.pciu_email
+          'veteranAddress' => address,
+          'email' => user.va_profile_email
         }
       end
 
@@ -1353,7 +1166,7 @@ RSpec.describe FormProfile, type: :model do
         before do
           allow(user).to receive(:icn).and_return('1012829228V424035')
           allow(Flipper).to receive(:enabled?).and_call_original
-          allow(Flipper).to receive(:enabled?).with(:remove_pciu, anything).and_return(false)
+          allow(Flipper).to receive(:enabled?).with(:remove_pciu, anything).and_return(true)
         end
 
         context "when the 'ezr_form_prefill_with_providers_and_dependents' flipper is enabled" do
@@ -1522,7 +1335,7 @@ RSpec.describe FormProfile, type: :model do
                 'veteranFullName' => full_name,
                 'address' => address,
                 'telephoneNumber' => us_phone,
-                'emailAddress' => user.pciu_email,
+                'emailAddress' => user.va_profile_email,
                 'dateOfBirth' => user.birth_date
               },
               'income' => [
@@ -1582,35 +1395,14 @@ RSpec.describe FormProfile, type: :model do
       end
     end
 
-    context 'user without an address' do
-      it 'prefills properly' do
-        expect(user).to receive(:address).exactly(6).times.and_return(
-          street: nil,
-          street2: nil,
-          city: nil,
-          state: nil,
-          country: nil,
-          postal_code: nil
-        )
-        described_class.for(form_id: '22-1990e', user:).prefill
-      end
-    end
-
     context 'with military information data', :skip_va_profile do
       context 'with va profile prefill on' do
         before do
           VAProfile::Configuration::SETTINGS.prefill = true
-
-          v22_1990_expected['email'] = VAProfileRedis::ContactInformation.for_user(user).email.email_address
+          v22_1990_expected['email'] = VAProfileRedis::V2::ContactInformation.for_user(user).email.email_address
           v22_1990_expected['homePhone'] = '3035551234'
           v22_1990_expected['mobilePhone'] = '3035551234'
-          v22_1990_expected['veteranAddress'] = {
-            'street' => '140 Rock Creek Rd',
-            'city' => 'Washington',
-            'state' => 'DC',
-            'country' => 'USA',
-            'postalCode' => '20011'
-          }
+          v22_1990_expected['veteranAddress'] = address
         end
 
         after do
@@ -1654,7 +1446,7 @@ RSpec.describe FormProfile, type: :model do
         end
 
         it 'prefills 0994 with VA Profile and payment information' do
-          VCR.use_cassette('evss/pciu_address/address_domestic') do
+          VCR.use_cassette('va_profile/v2/contact_information/get_address') do
             VCR.use_cassette('evss/disability_compensation_form/rated_disabilities') do
               VCR.use_cassette('evss/ppiu/payment_information') do
                 VCR.use_cassette('va_profile/military_personnel/post_read_service_histories_200',
@@ -1729,7 +1521,7 @@ RSpec.describe FormProfile, type: :model do
                 'serviceNumber' => '123455678'
               },
               'contactInformation' => {
-                'email' => user.pciu_email,
+                'email' => user.va_profile_email,
                 'phone' => us_phone,
                 'workPhone' => '13035551234',
                 'address' => address
@@ -1755,8 +1547,8 @@ RSpec.describe FormProfile, type: :model do
 
       context 'with VA Profile prefill for 10203' do
         before do
-          can_prefill_vaprofile(true)
           expect(user).to receive(:authorize).with(:evss, :access?).and_return(true).at_least(:once)
+          expect(user).to receive(:authorize).with(:va_profile, :access?).and_return(true).at_least(:once)
         end
 
         it 'prefills 10203' do
@@ -1782,7 +1574,7 @@ RSpec.describe FormProfile, type: :model do
         end
 
         it 'prefills 10203 with VA Profile and entitlement information' do
-          VCR.use_cassette('evss/pciu_address/address_domestic') do
+          VCR.use_cassette('va_profile/v2/contact_information/get_address') do
             VCR.use_cassette('evss/disability_compensation_form/rated_disabilities') do
               VCR.use_cassette('form_10203/gi_bill_status_200_response') do
                 VCR.use_cassette('gi_client/gets_the_institution_details') do
@@ -1808,27 +1600,21 @@ RSpec.describe FormProfile, type: :model do
           can_prefill_vaprofile(true)
         end
 
-        context 'with a user with no vet360_id' do
-          before do
-            allow(user).to receive(:vet360_id).and_return(nil)
+        it 'omits address fields in 686c-674 form' do
+          VCR.use_cassette('va_profile/military_personnel/post_read_service_histories_200',
+                           allow_playback_repeats: true) do
+            prefilled_data = described_class.for(form_id: '686C-674', user:).prefill[:form_data]
+            v686_c_674_expected['veteranContactInformation']
+            expect(prefilled_data).to eq(v686_c_674_expected)
           end
+        end
 
-          it 'omits address fields in 686c-674 form' do
-            VCR.use_cassette('va_profile/military_personnel/post_read_service_histories_200',
-                             allow_playback_repeats: true) do
-              prefilled_data = described_class.for(form_id: '686C-674', user:).prefill[:form_data]
-              v686_c_674_expected['veteranContactInformation'].delete('veteranAddress')
-              expect(prefilled_data).to eq(v686_c_674_expected)
-            end
-          end
-
-          it 'omits address fields in 686c-674-V2 form' do
-            VCR.use_cassette('va_profile/military_personnel/post_read_service_histories_200',
-                             allow_playback_repeats: true) do
-              prefilled_data = described_class.for(form_id: '686C-674-V2', user:).prefill[:form_data]
-              v686_c_674_expected_v2['veteranContactInformation'].delete('veteranAddress')
-              expect(prefilled_data).to eq(v686_c_674_expected_v2)
-            end
+        it 'omits address fields in 686c-674-V2 form' do
+          VCR.use_cassette('va_profile/military_personnel/post_read_service_histories_200',
+                           allow_playback_repeats: true) do
+            prefilled_data = described_class.for(form_id: '686C-674-V2', user:).prefill[:form_data]
+            v686_c_674_expected_v2['veteranContactInformation']
+            expect(prefilled_data).to eq(v686_c_674_expected_v2)
           end
         end
 
@@ -1854,6 +1640,7 @@ RSpec.describe FormProfile, type: :model do
           FORM-MOCK-AE-DESIGN-PATTERNS
         ].each do |form_id|
           it "returns prefilled #{form_id}" do
+            allow(Flipper).to receive(:enabled?).with(:pension_military_prefill, anything).and_return(false)
             VCR.use_cassette('va_profile/military_personnel/service_history_200_many_episodes',
                              allow_playback_repeats: true, match_requests_on: %i[uri method body]) do
               expect_prefilled(form_id)
@@ -1861,73 +1648,36 @@ RSpec.describe FormProfile, type: :model do
           end
         end
 
-        context 'with a user that can prefill' do
+        context 'when Vet360 prefill is enabled' do
+          let(:user) do
+            build(:user, :loa3, :legacy_icn, suffix: 'Jr.', address: build(:va_profile_v3_address),
+                                             vet360_id: '1781151')
+          end
+
           before do
-            allow_any_instance_of(Auth::ClientCredentials::Service).to receive(:get_token).and_return('usyergd')
+            VAProfile::Configuration::SETTINGS.prefill = true # TODO: - is this missing in the failures above?
+            expected_veteran_info = v21_526_ez_expected['veteran']
+            expected_veteran_info['emailAddress'] = user.va_profile_email
+            expected_veteran_info['primaryPhone'] = us_phone
+            allow_any_instance_of(Auth::ClientCredentials::Service).to receive(:get_token).and_return('fake_token')
           end
 
-          # NOTE: `increase only` and `all claims` use the same form prefilling
-          context 'when Vet360 prefill is disabled' do
-            let(:user) do
-              build(:user, :loa3, :legacy_icn, suffix: 'Jr.', address: build(:mpi_profile_address))
-            end
-
-            before do
-              expect(user).to receive(:authorize).with(:lighthouse, :direct_deposit_access?)
-                                                 .and_return(true).at_least(:once)
-              expect(user).to receive(:authorize).with(:evss, :access?).and_return(true).at_least(:once)
-            end
-
-            it 'returns prefilled 21-526EZ' do
-              VCR.use_cassette('evss/pciu_address/address_domestic') do
-                VCR.use_cassette('lighthouse/veteran_verification/disability_rating/200_response') do
-                  VCR.use_cassette('lighthouse/direct_deposit/show/200_valid_new_icn') do
-                    VCR.use_cassette('va_profile/military_personnel/service_history_200_many_episodes',
-                                     allow_playback_repeats: true, match_requests_on: %i[uri method body]) do
-                      VCR.use_cassette('disability_max_ratings/max_ratings') do
-                        expect_prefilled('21-526EZ')
-                      end
-                    end
-                  end
-                end
-              end
-            end
+          after do
+            VAProfile::Configuration::SETTINGS.prefill = false
           end
-        end
 
-        context 'without Lighthouse direct deposit' do
-          context 'when Vet360 prefill is enabled' do
-            let(:user) do
-              build(:user, :loa3, :legacy_icn, suffix: 'Jr.', address: build(:mpi_profile_address))
-            end
-
-            before do
-              VAProfile::Configuration::SETTINGS.prefill = true # TODO: - is this missing in the failures above?
-              expected_veteran_info = v21_526_ez_expected['veteran']
-              expected_veteran_info['emailAddress'] =
-                VAProfileRedis::V2::ContactInformation.for_user(user).email.email_address
-              expected_veteran_info['primaryPhone'] = '3035551234'
-              allow_any_instance_of(Auth::ClientCredentials::Service).to receive(:get_token).and_return('fake_token')
-              allow(Flipper).to receive(:enabled?).with(:remove_pciu, anything).and_return(true)
-            end
-
-            after do
-              VAProfile::Configuration::SETTINGS.prefill = false
-            end
-
-            it 'returns prefilled 21-526EZ' do
-              expect(user).to receive(:authorize).with(:lighthouse, :direct_deposit_access?)
-                                                 .and_return(true).at_least(:once)
-              expect(user).to receive(:authorize).with(:evss, :access?).and_return(true).at_least(:once)
-              expect(user).to receive(:authorize).with(:va_profile, :access_to_v2?).and_return(true).at_least(:once)
-              VCR.use_cassette('evss/pciu_address/address_domestic') do
-                VCR.use_cassette('lighthouse/veteran_verification/disability_rating/200_response') do
-                  VCR.use_cassette('lighthouse/direct_deposit/show/200_valid_new_icn') do
-                    VCR.use_cassette('va_profile/military_personnel/service_history_200_many_episodes',
-                                     allow_playback_repeats: true, match_requests_on: %i[uri method body]) do
-                      VCR.use_cassette('disability_max_ratings/max_ratings') do
-                        expect_prefilled('21-526EZ')
-                      end
+          it 'returns prefilled 21-526EZ' do
+            expect(user).to receive(:authorize).with(:lighthouse, :direct_deposit_access?)
+                                               .and_return(true).at_least(:once)
+            expect(user).to receive(:authorize).with(:evss, :access?).and_return(true).at_least(:once)
+            expect(user).to receive(:authorize).with(:va_profile, :access_to_v2?).and_return(true).at_least(:once)
+            VCR.use_cassette('va_profile/v2/contact_information/get_address') do
+              VCR.use_cassette('lighthouse/veteran_verification/disability_rating/200_response') do
+                VCR.use_cassette('lighthouse/direct_deposit/show/200_valid_new_icn') do
+                  VCR.use_cassette('va_profile/military_personnel/service_history_200_many_episodes',
+                                   allow_playback_repeats: true, match_requests_on: %i[uri method body]) do
+                    VCR.use_cassette('disability_max_ratings/max_ratings') do
+                      expect_prefilled('21-526EZ')
                     end
                   end
                 end
@@ -1977,7 +1727,7 @@ RSpec.describe FormProfile, type: :model do
         allow_any_instance_of(BGS::People::Service).to(
           receive(:find_person_by_participant_id).and_return(BGS::People::Response.new({ file_nbr: '1234567890' }))
         )
-        allow_any_instance_of(VAProfile::Models::Address).to(
+        allow_any_instance_of(VAProfile::Models::V3::Address).to(
           receive(:address_line3).and_return('suite 500')
         )
       end
@@ -2016,7 +1766,7 @@ RSpec.describe FormProfile, type: :model do
         allow_any_instance_of(BGS::People::Service).to(
           receive(:find_person_by_participant_id).and_return(BGS::People::Response.new({ file_nbr: '1234567890' }))
         )
-        allow_any_instance_of(VAProfile::Models::Address).to(
+        allow_any_instance_of(VAProfile::Models::V3::Address).to(
           receive(:address_line3).and_return('suite 500')
         )
       end
